+++ packages/react-form-with-constraints/src/Async.tsx
@@ -87,7 +87,7 @@ export class Async<T> extends
 
     const field = form.fieldsStore.getField(input.name)!;
 
-    if (fieldFeedbacks.props.stop === 'first' && field.hasAnyFeedbacks(fieldFeedbacks.key) ||
+    if (fieldFeedbacks.props.stop === 'first' && field.hasFeedbacks(fieldFeedbacks.key) ||
         fieldFeedbacks.props.stop === 'first-error' && field.hasErrors(fieldFeedbacks.key) ||
         fieldFeedbacks.props.stop === 'first-warning' && field.hasWarnings(fieldFeedbacks.key) ||
         fieldFeedbacks.props.stop === 'first-info' && field.hasInfos(fieldFeedbacks.key)) {